Introduce DisableInHyperlinks and walk markdown tree for each link
New test cases added covering link insertion and shortening for
JIRA-ticket patterns.

A pattern that is not intended ever to match a hyperlink can now be
marked as such, and can therefore use WordMatch without matching part of
a hyperlink.

Since one pattern may insert a new hyperlink and we can't let a later
pattern match against it, we also walk the markdown tree for each link
(based on latest modified content) instead of applying each link to each
node of the markdown tree.