Correctly handle updated posts
If the first return value of MessageWillBeUpdated is nil, the update is
rejected, with the second return value providing an explanation for the
rejection. To allow the update to occur without any changes, the
function should return the value of newPost that was passed in to it.

We don't reject any post updates, but if we did then even then we would
have to provide a rejection reason. As such, we don't need to alter the
logic of ProcessPost, and can instead simply put a check on its return
values.